moved scripts
idk what to do man, theres so many damn merge errors and i literally have no idea what the hell to do about it. i dont know what changes were made, i didnt make them! i only waited for so long to start programming because i thought i would simply be able to start adding on to what kevin and jordy finished, but all of kevins work is giving some insane merge errors for various reasons. partially because a lot of the changes were made outside of the actual snake prefab and just on the scene itself. also he was working off of an older version of main for that, so after i made more changes it gives a ton of merge errors when i try and put kevins work in. god damn it man, i dont know how to fix this shit. i just wanted to finish the powerups and put in the dash tonight, but if i want to do anything, then i need to fix all this bullshit. i might just go to sleep and say fuck it ill do it tommorow i clearly dont have any motivation to fix all this bullshit tonight. i guess if i made the powerup it would be find but actually no it wouldnt because i would need to make changes to the player and enemy scripts for that and that would literally only cause a ton of more fucking merge errors.

https://pbs.twimg.com/media/GHDX6qPawAAZUKV?format=jpg&name=medium